Completato Selfwork Broken Access Control

+++ app/Http/Controllers/ArticleController.php
@@ -86,6 +86,10 @@ class ArticleController extends Controller
 
     public function edit(Article $article)
     {
+        // SECURE
+        if(!Auth::id() != $article->user_id && !Auth::user()->isAdmin()){
+            return redirect()->route('articles.index', $article)->with('message','Not authorized');
+        }
         return view('articles.edit',compact('article'));
     }
 
@@ -109,9 +113,9 @@ class ArticleController extends Controller
     public function destroy(Article $article, Request $request)
     {
         // SECURE
-        // if(Auth::id() !== $article->user_id){
-        //     return redirect()->route('articles.show', $article)->with('message','Not authorized');
-        // }
+        if(Auth::id() !== $article->user_id){
+            return redirect()->route('articles.show', $article)->with('message','Not authorized');
+        }
         
         $article->delete();
